# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Admin panel API for a Dune Awakening private server.",
        "title": "dune-admin API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/blueprints": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "blueprints"
                ],
                "summary": "List all building blueprints",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "additionalProperties": true
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/blueprints/import": {
            "post": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "blueprints"
                ],
                "summary": "Import a blueprint JSON file into a player's inventory",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Blueprint JSON file",
                        "name": "blueprint",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Player pawn ID to receive the blueprint",
                        "name": "player_id",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/blueprints/{id}/export": {
            "get": {
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "blueprints"
                ],
                "summary": "Export a blueprint as a downloadable JSON file",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Blueprint ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Blueprint JSON file",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}